added error handling and minor refactor
