+++ src/errors/NeoConnectionError.ts
