+++ src/modules/feed-querier/feed-query-config.ts
@@ -19,28 +19,28 @@ export const CALCULATE_LIKELY_BITE_CANDIDATES = `ORDER BY bite.score ${ORDER_BY_
 export const MATCH_INTERACTED_USER_CHAIN =
     'MATCH ' +
     '(user)-[interaction:InteractsWith]->(authoredUser:User)-[:Authored]->' +
-    '(post:Post)-[posted:PostedIn]->(thread:Thread)<-[:HasThread]-(bite:Bite)';
+    '(post)-[posted]->(thread:Thread)<-[:HasThread]-(bite:Bite)';
 
 //communities that have been interacted with by the target user
 export const MATCH_INTERACTED_BITE_CHAIN =
     'MATCH ' +
     '(user)-[interaction:InteractsWith]->(bite:Bite)-[:HasThread]->' +
-    '(thread:Thread)<-[posted:PostedIn]-(post:Post)<-[:Authored]-(authoredUser:User)';
+    '(thread:Thread)<-[posted]-(post)<-[:Authored]-(authoredUser:User)';
 
 //similar communities based on common followers
 export const MATCH_SIMILAR_BITE_CHAIN =
     'MATCH ' +
     '(user)-[:InteractsWith]->(:Bite)-[interaction:SimilarBite]->(bite:Bite)' +
-    '-[:HasThread]->(thread:Thread)<-[posted:PostedIn]-(post:Post)<-[:Authored]-(authoredUser:User)';
+    '-[:HasThread]->(thread:Thread)<-[posted]-(post)<-[:Authored]-(authoredUser:User)';
 
 //mutual users posts (follower of follower)
 export const MATCH_MUTUAL_USER_CHAIN =
     'MATCH ' +
     '(user)-[interaction:SimilarMutual]->(authoredUser:User)-[:Authored]->' +
-    '(post:Post)-[posted:PostedIn]->(thread:Thread)<-[:HasThread]-(bite:Bite)';
+    '(post)-[posted]->(thread:Thread)<-[:HasThread]-(bite:Bite)';
 
 //mutual bite posts (community followed by followed user)
 export const MATCH_MUTUAL_BITE_CHAIN =
     'MATCH ' +
     '(user)-[interaction:SimilarMutual]->(bite:Bite)-[:HasThread]->' +
-    '(thread:Thread)<-[posted:PostedIn]-(post:Post)<-[:Authored]-(authoredUser:User)';
+    '(thread:Thread)<-[posted]-(post)<-[:Authored]-(authoredUser:User)';
